Lists: * Move List objects and functions from Numbers module to Lists module
