+++ lists/src/main/kotlin/mathtools/lists/IntList.kt
@@ -1,4 +1,4 @@
-package mathtools.numbers.listtypes
+package mathtools.lists
 
 import java.math.BigInteger
 import kotlin.math.roundToLong
